requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.1
Automat==22.10.0
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.1
daphne==4.0.0
defusedxml==0.7.1
Django==4.2.2
django-celery-beat==2.5.0
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-filter==23.2
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dnspython==2.3.0
flower==1.2.0
gunicorn==20.1.0
h11==0.14.0
humanize==4.6.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.0
Markdown==3.4.3
oauthlib==3.2.2
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pykerberos==1.2.4
pyOpenSSL==23.2.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-snappy==0.6.1
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
sentry-sdk==1.25.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tornado==6.3.2
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.0
